\name{automaticInputList}

\alias{automaticInputList}

\title{
Automatic input list.
}

\description{
Automatic input list.
}

\usage{
gm <- Gelito()
gm$automaticInputList(id = NA, nf = list(), cc = list(), cn = list())
}

\arguments{
  \item{id}{
ID column number.
}
  \item{nf}{
filter columns numbers.
}
  \item{cc}{
criterions.
}
  \item{cn}{
column numbers.
}
}

\value{
no
}

\author{
Manuel Ramón Gutiérrez Covarrubias
}

\seealso{
Gelito
}
