\name{common}

\alias{common}

\title{
Looking for shared elements in lists.
}

\description{
Looking for shared elements in lists.
}

\usage{
gm <- Gelito()
gm$common(lists = list(), ncol = 1)
}

\arguments{
  \item{lists}{
data lists.
}
  \item{ncol}{
sharing will be done by number of ncol.
}
}

\value{
  \item{list}{
list result
}
}

\author{
Manuel Ramón Gutiérrez Covarrubias
}

\seealso{
Gelito
}
