\name{intersectVenn}

\alias{intersectVenn}

\title{
Venn intersect.
}

\description{
Venn intersect. Draw diagram.
}

\usage{
gm <- Gelito()
gm$intersectVenn(lists = list(), ncol = 1, names = c())
}

\arguments{
  \item{lists}{
list of number of list to intersect; count must be greater than 1.
}
  \item{ncol}{
intersect will be done by number of ncol.
}
  \item{names}{
columns names to show.
}
}

\value{
Venn diagram.
}

\author{
Manuel Ramón Gutiérrez Covarrubias
}

\seealso{
Gelito
}
