\name{manualInputList}

\alias{manualInputList}

\title{
Manual input list.
}

\description{
Environment is configured by console.
}

\usage{
gm <- Gelito()
gm$manualInputList(filename = NA)
}

\arguments{
  \item{filename}{
output filename.
}
}

\value{
no
}

\author{
Manuel Ramón Gutiérrez Covarrubias
}

\seealso{
Gelito
}
