\name{select}

\alias{select}

\title{
Returns lists with columns value.
}

\description{
Returns lists with columns value. Select one column from list, and return it.
}

\usage{
gm <- Gelito()
gm$select(lists = list(), ncol = 1)
}

\arguments{
  \item{lists}{
data lists.
}
  \item{format}{
number of column to create list.
}
}

\value{
  \item{list}{
list with one column.
}
}

\author{
Manuel Ramón Gutiérrez Covarrubias
}

\seealso{
Gelito
}
