\name{setMatrixFromFile}

\alias{setMatrixFromFile}

\title{
Read expression matrix from file.
}

\description{
Read expression matrix from file and set "em" attribute.
}

\usage{
gm <- Gelito()
gm$setMatrixFromFile(filename = NA)
}

\arguments{
  \item{filename}{
filename with expression matrix.
}
}

\value{
no
}

\author{
Manuel Ramón Gutiérrez Covarrubias
}

\seealso{
Gelito
}
