\name{subset}

\alias{subset}

\title{
Subsetting expression matrix.
}

\description{
Subsetting expression matrix with list.
}

\usage{
gm <- Gelito()
gm$subset(vector = c())
}

\arguments{
  \item{vector}{
vector for subsetting.
}
}

\value{
  \item{data.frame}{
subset result
}
}

\author{
Manuel Ramón Gutiérrez Covarrubias
}

\seealso{
Gelito
}
