\name{union}

\alias{union}

\title{
List Union.
}

\description{
List Union.
}

\usage{
gm <- Gelito()
gm$union(lists = list(), ncol = 1)
}

\arguments{
  \item{lists}{
list of number of list to union; count must be greater than 0.
}
  \item{ncol}{
union will be done by number of ncol.
}
}

\value{
  \item{list}{
list result
}
}

\author{
Manuel Ramón Gutiérrez Covarrubias
}

\seealso{
Gelito
}
