\name{update}

\alias{update}

\title{
Update creator data attribute.
}

\description{
Update creator data attribute like data frame.
}

\usage{
gm <- Gelito()
gm$update(data = NA)
}

\arguments{
  \item{data}{
data frame with subset. It can be a list too.
}
}

\value{
no
}

\author{
Manuel Ramón Gutiérrez Covarrubias
}

\seealso{
Gelito
}
