\name{writeFile}

\alias{writeFile}

\title{
Write file.
}

\description{
Write file with format.
}

\usage{
gm <- Gelito()
gm$writeFile(filename = NA, format = NA)
}

\arguments{
  \item{filename}{
output filename.
}
  \item{format}{
output file format (txt, csv or xls).
}
}

\value{
no
}

\author{
Manuel Ramón Gutiérrez Covarrubias
}

\seealso{
Gelito
}
